+++ miden/tests/integration/operations/io_ops/mem_ops.rs
@@ -1,3 +1,6 @@
+use prover::ExecutionError;
+use test_utils::expect_exec_error;
+
 use super::{apply_permutation, build_op_test, build_test, Felt, ToElements, TRUNCATE_STACK_PROC};
 
 // LOADING SINGLE ELEMENT ONTO THE STACK (MLOAD)
@@ -227,4 +230,25 @@ fn read_after_write() {
     // --- write to memory first, then test read with loadw --------------------------------------
     let test = build_op_test!("mem_storew.0 dropw mem_loadw.0", &[1, 2, 3, 4, 5, 6, 7, 8]);
     test.expect_stack(&[8, 7, 6, 5]);
+}
+
+// MISC
+// ================================================================================================
+
+/// Ensures that the processor returns an error when 2 memory operations occur in the same context,
+/// at the same address, and in the same clock cycle (which is what RCOMBBASE does when `stack[13] =
+/// stack[14] = 0`).
+#[test]
+fn mem_reads_same_clock_cycle() {
+    let asm_op = "begin rcomb_base end";
+
+    let test = build_test!(asm_op);
+    expect_exec_error!(
+        test,
+        ExecutionError::DuplicateMemoryAccess {
+            ctx: 0_u32.into(),
+            addr: 0,
+            clk: 1_u32.into()
+        }
+    );
 }